anomaly {multivariateAnomaly}	R Documentation

Anomaly Detection given a non-negative symmetrix Kernel matrices.

Description

This function outputs the connectivity values between the vertices of the Kernel. Lower the value, more anomalous the vertex in the original input.

Usage

anomaly(KX, KY, d)

Arguments

KX

KY

d

Author(s)

Sriram Lakshminarasimhan Siddarth Ramaswamy

References

[1] Haibin Cheng, Pang-Ning Tan, Christopher Potter, and Steven Kl A Robust Graph-based Algorithm for Detection and Characterization Multivariate Time Series, to appear in Proc.of IEEE International workshop on Spatial and Spatio-temporal Data Mining (ICDM/STDM 08)

[Package multivariateAnomaly version 1.0 Index]