distance {multivariateAnomaly}	R Documentation

Calculate time series similarity distance measure.

Description

Given a set of measurements and their timestamps, this function outputs a symmetric matrix that capture the pair wise similarity between every pair of timestamps.

Usage

distance(X)

Arguments

X 	 X is a matrix, consisting of $p$ real valued measurements in each column (timestamp).

Value

A non-negative symmetric matrix is returned.

Author(s)

Sriram Lakshminarasimhan Siddarth Ramaswamy

References

Haibin Cheng, Pang-Ning Tan, Christopher Potter, Steven Klooster, "A Robust Graph-Based Algorithm for Detection and Characterization of Anomalies in Noisy Multivariate Time Series," icdmw, pp.349-358, 2008 IEEE International Conference on Data Mining Workshops, 2008

See Also

Other random walk algorithms 'outlierA', 'outlierB'

Examples

X <- matrix(c(1:9), nrow=3, ncol=3, byrow=TRUE);
distance(X)

[Package multivariateAnomaly version 1.0 Index]