matrixAlign {multivariateAnomaly}	R Documentation

Alignment of matrices.

Description

The alignMatrix, takes in two kernel similarity matrices as input and aligns them i.e. it maximizes the correlation between the two matrices.

Usage

matrixAlign(KX, KY)

Arguments

X 	 Kernel alignment of the predictor time series variables.

Y 	 Kernel alignment of the target variables.

Author(s)

Sriram Lakshminarasimhan

References

[1] Haibin Cheng, Pang-Ning Tan, Christopher Potter, and Steven Klooster. A Robust Graph-based Algorithm for Detection and Characterization of Anomalies in Noisy Multivariate Time Series, to appear in Proc.of IEEE International Conference on Data Mining workshop on Spatial and Spatio-temporal Data Mining (ICDM/STDM 08)

Examples

data(KernelPredictorSample1)
data(KernelTargetSample1)
X = as.matrix(KernelPredictorSample1)
Y = as.matrix(KernelTargetSample1)
X
KX = RBFMeasure(X, 1.414)
KY = RBFMeasure(Y, 0.44)
K_Aligned = matrixAlign(KX, KY)
K_Aligned

[Package multivariateAnomaly version 1.0 Index]