multivariateAnomaly {multivariateAnomaly}	R Documentation

Anomaly detection in time-series measurements

Description

Performs anomaly detection using algorithm descriped in the referenced paper, for a set of multivariate time series predictor values and a set of target values.

Usage

multivariateAnomaly(X, Y, r, sig, nv, d = 0.1)

Arguments

X 	 The predictor time series variables, each row is a variable and each column is a timestamp.

Y 	 The target time series.

r 	 The autoregression lag.

sig 	 The spread width of the RBF kernel.

nv 	 Number of eigenvectors to be used.

d 	 The damping factor to be passed to the anomaly detection function. Default value is 0.1

Value

Intermediate outputs produced in this step passed to anomaly detection function

w 	 The weight associated with each eigenvector.

KX 	 The new aligned kernel matrix for X.

KY 	 The kernel matrix for target Y.

Author(s)

Sriram Lakshminarasimhan Siddarth Ramaswamy

References

[1] Haibin Cheng, Pang-Ning Tan, Christopher Potter, and Steven Klooster. A Robust Graph-based Algorithm for Detection and Characterization of Anomalies in Noisy Multivariate Time Series, to appear in Proc.of IEEE International Conference on Data Mining workshop on Spatial and Spatio-temporal Data Mining (ICDM/STDM 08)

Examples

# X is the predictor, having 2 variables whose values have been measured from
# t1 to t5 . Y is the target variable.
data(KernelPredictorSample2)
data(KernelTargetSample2)
X = as.matrix(KernelPredictorSample2)
Y = as.matrix(KernelTargetSample2)
X
Y
# multivariateAnomaly function takes in input target and predictor matrices, 
# along with the number of eigen values and the spread width of the RBF kernel.
# Damping factor is an optional parameter.
RandomWalkOutput = multivariateAnomaly(X, Y, 0, 93.14, 5)
RandomWalkOutput

[Package multivariateAnomaly version 1.0 Index]